Update README.md
1. Among Us

In the game of Among Us, there are 9 players. 4 of them are imposters and 5 of them are crewmates. There is also a deck of 17 cards containing 11 \sabotage” cards and 6 \task” cards. Imposters want to play sabotage cards, and crewmates want to play task cards. Here’s how the play proceeds.

A captain and a rst mate are chosen uniformly at random from the 9 players.

The captain draws 3 cards from the deck and gives 2 to the rst mate, discarding the third.

The rst mate chooses one to play.

Now suppose you are the rst mate, but the captain gave you 2 sabotage cards. Being a crewmate, you wonder, did the captain just happen to have 3 sabotage cards, or was the captain an imposter who secretly discarded a task card. In this scenario, what’s the probability that the captain is an imposter? Let’s assume that imposter captains always try to discard task cards, and crewmate captains always try to discard sabotage cards.

2. Lightbulbs